+++ AI/src/Testsuites/tensortest.cpp
@@ -120,7 +120,7 @@ int convertTensortestCircuitToTikz(int index) {
           quake_source_input_file)) {
     return -1;
   }
-  InstructionsTensor<double> observation =
+  InstructionsTensor<float> observation =
       quantum_circuit_environment.get_observation();
   QuantumCircuit circuit = recreateQuantumCircuitFromTensor(observation);
   if (int rc = write_to_file(&circuit, latex_quake_output_file); rc != 0) {
